Fixed landscape UI causing reboots.
Here's the crash details.
Open an app that you know for sure is going to cause the phone to go into landscape mode, for my testing I use Dolphin Browser.
Now, with the phone in landscape mode, lock the phone with the side button. Press the lock button to try and unlock it, it won't unlock.
Press and hold the unlock button and now notice that your phone is rebooting.

code by trparky on irc (thomas.parkison@gmail.com)

Change-Id: I3e6c37b06c130ca81b5fc0c159e98a7b7bf36727
Signed-off-by: iceandfire <arhamjamal@gmail.com>