Lockscreen: enable longpress on expand challenge handle (1/2)
If enabled it will be possible to longpress the little lock icon.

If it is an unsecure screen lock then the longpress will unlock the device.
If it is secure then the widget will fade into the background and you need
to enter the pattern, pin, password or whatever.

Change-Id: I901424c8c54d59298d6d77c10154f5410a636b01
Signed-off-by: Dirk Rettschlag <dirk.rettschlag@gmail.com>
Signed-off-by: iceandfire <arhamjamal@gmail.com>

Conflicts:
	core/java/android/provider/Settings.java
	core/res/res/values/symbols.xml
	policy/src/com/android/internal/policy/impl/keyguard/KeyguardHostView.java

Conflicts:
	core/java/android/provider/Settings.java