FW : Rework NavBar UI in ROMControl.
This turned out to be a little more in depth than I planned.
Decided to do some integration with AwesomeAction while
overhauling the UI.  This required adding some methods to
AwesomeConstants so that we can get the list of actions and
their 'ProperNames'  from a single object.

ToDo:  Do a similar move of objects/names for the icons.
   This will allow a single copy of each icon rather than
two like now.

Themers and Translaters are gonna be pissed when this is done.

PS2:  Created a method to return the Drawable/Icon associated
with a given action.  I've imported the XHDPI icons only at this
time (so non-xhdpi test builds will not work!)
Also, some of our 'actions' didn't have associated icons so I
copied a few from other areas.  They will need to be looked at
for 'appropriateness'.

ps3:  Moved NavBarHelpers into frameworks.  Copied over the rest
of the DPI icons (Resized quite a few too) - the images still need
review/updating to be more holified, but they will work as is.
Todo: started a way to subract subsets of Actions for the appropriate
function (ie, NavBar might not need access to *all* Actions.  I didn't
implement yet as I need a clever/clean way to subtract Array sets.

ps4:  Had to remove a few references to moved icons in SystemUI

ps5: Moved Icons back to SystemUI made a SystemUIResource getter method.
     Made navring use navbar icons.

ps6: update a few icons - thanks to Nitroz

Change-Id: Ia58f25f600fb92b70a4658e16a512e9a9ea8430f

Conflicts:
	core/java/com/android/internal/util/aokp/AwesomeConstants.java
	core/res/res/values/strings.xml
	core/res/res/values/symbols.xml
	packages/SystemUI/res/drawable-xhdpi/ic_sysbar_back_land.png
	packages/SystemUI/res/drawable-xhdpi/ic_sysbar_home_land.png
	packages/SystemUI/res/drawable-xhdpi/ic_sysbar_recent_land.png
	packages/SystemUI/src/com/android/systemui/statusbar/NavigationBarView.java

Signed-off-by: iceandfire <arhamjamal@gmail.com>