P2P: Fix the failure of cancel P2P invitation.
Framework will simply reply a CANCEL_CONNECT_FAILED with reason that
framework is in BUSY when try to cancel a new invitation if there is
an already exsisted p2p connection. It will cause the failure of
cancel P2P invitation. So change the behaviour here to reply
CANCEL_CONNECT_SUCCEEDED and cancel the invitation.

Change-Id: I9b270599120d2ea12dea755613691398568b37b5