Dynamic glowScale adjustment and spacing of NavBar.
This adjust the Glowscale for the NavBar buttons based on
the number of buttons on the bar.  I also added spacer
views to the outer edges of the NavBar.  This helps alleviate
conditions where the glow was cut off on outer buttons (especially
in tablet UI).  It also makes the Glow more appropriately sized
as the number of buttons increase.

There are still some cases where an excess number of buttons,
and a small navbar percentage (TabUI) is going to create some
glow cut-off conditions.  That's just going to be unavoidable when
the user tries to squeeze it down that small.

ps 2: Adding the spacers now only occurs on TabUI.  This prevents
    major changes to the look/layout of buttons on Phone/Phablet,
    but still provides glow cutoff release for tablets, as they
    were the worst effected.
ps3 : adjust the size of the separators to not squeeze navbar too much

PS4 : whitespaces

Change-Id: Ic6d9deb4ae49c87251b8b258ff0f42da46910986
Signed-off-by: Zaphod <mwilson@acesappraisal.com>
Signed-off-by: iceandfire <arhamjamal@gmail.com>

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/NavigationBarView.java