multiwindow/display apps in dialogs
first draft. after much trying i found the one single combination of flags that
makes it possible without trashing the actionbar or getting crashes.
its hardcoded for mms, chrome, gmail, gtalk and recents. --> for testing ofc

we need to track down why it blacks out the background. a slight dim and we're
pretty much done - then we just need to add intelligent triggers to make it
happen only for notifications.

Change-Id: I18f1eab3ad199204df384fdb7e7ad8a709e1f7fe