Active display: Enable proximity sensor only when screen is off
[log]

This should ease people worried about cpu usage of the proximity
sensor while the screen is on.  The sensor listener will only be
registered when the screen goes off, and both active display and
pocket mode are enabled.  Once the screen turns on this listener
will be unregistered so no further updates are received when the
sensor values change.

by Der_Schubi:
In our case we left the sensor activated til the screen gets
unlocked or notification is opened to be able to turn the screen
back off if the device gets pocketed again.

Change-Id: Ic28326cc0cb0fb87d4bc0d1e3224e04727242751

Conflicts:
	packages/SystemUI/src/com/android/systemui/statusbar/policy/activedisplay/ActiveDisplayView.java